+++ Sources/CoreData/Database+Operation.swift
@@ -17,19 +17,20 @@ public extension DB {
 
 	/// Perform an operation asynchronously.
 	///
+	/// - parameter queue: The queue on which your completion handler is dispatched.
 	/// - parameter operation: The closure to perform within a new context.
 	/// - parameter context: The temporary save context.
 	/// - parameter save: The save closure, you should call this when you're ready to save.
-	func operation(_ operation: @escaping (_ context: NSManagedObjectContext, _ save: @escaping SaveBlockWitCallback) -> Void) {
+	func operation(queue: DispatchQueue = .main, operation: @escaping (_ context: NSManagedObjectContext, _ save: @escaping SaveBlockWitCallback) -> Void) {
 		container.performBackgroundTask { context in
 			context.mergePolicy = NSMergePolicy(merge: self.mergePolicy)
 
 			operation(context) { callback in
 				do {
 					try context.save()
-					callback(nil)
+					queue.async { callback(nil) }
 				} catch {
-					callback(error)
+					queue.async { callback(error) }
 				}
 			}
 		}